add:修改TCP消息号，发送type=11,operation=16,接受是11,18

+++ src/main/english/java-en_tcp/com/xueersi/parentsmeeting/modules/livevideo/lib/TcpConstants.java
@@ -59,7 +59,7 @@ public class TcpConstants {
     /** roleplay分组使用的operation */
     public final static int ROLE_PLAY_INTERACTIVE_OPERATION = 2;
     /** roleplay自定义消息使用的type */
-    public final static int ROLE_PLAY_CUSTOM_TYPE = 11;
+    public final static short ROLE_PLAY_CUSTOM_TYPE = 11;
     /** roleplay自定义消息发送使用的operation */
     public final static int ROLE_PLAY_CUSTOM_OPERATION = 16;
     /** roleplay自定义消息接受使用的operation */
